+++ ogr/ogrsf_frmts/sqlite/ogrsqlitetablelayer.cpp
@@ -2793,8 +2793,9 @@ OGRErr OGRSQLiteTableLayer::BindValues(OGRFeature *poFeature,
                         poFieldDefn->GetSubType() == OFSTJSON)
                     {
                         json_object *poObjProp = nullptr;
-                        if (!OGRJSonParse(osValue.c_str(), &poObjProp, false,
-                                          osValue.length() + 1))
+                        if (!OGRJSonParse(
+                                osValue.c_str(), &poObjProp, false,
+                                static_cast<int>(osValue.length() + 1)))
                         {
                             // Emit warning once
                             CPLErrorOnce(
